feat(midi): idiomatic trait impls on the value types (From/TryFrom/Ord/Default)
Round out the public value types the way std::net's IP types are saturated with
conversions, so they flow through `.into()` and generic code instead of only
named constructors.

- NoteId: From<u32>/From<NoteId> for u32 (alongside from_raw/raw), and Ord so it
  works as a BTreeMap key / can be sorted.
- MidiUnitId: From<MidiUnitId> for u64 (completing the existing From<u64>), + Ord.
- MidiEvent: TryFrom<&[u8]> — the idiomatic byte-parse spelling alongside
  from_midi1_bytes (frame offset 0), with a typed MidiParseError.
- PitchBendSensitivity: Default (= MPE_DEFAULT, ±48 semitones per RP-053), a
  MPE_DEFAULT const, and Ord.
- Protocol: Default = Midi2 (this engine is MIDI-2-native).

Deliberately skipped the lossy/ambiguous ones: no From<MidiMessage> for MidiEvent
(a MidiMessage is a lossy view — re-encode would be a footgun), no Serialize
(Loro owns persistence), no Display on MidiEvent/MidiMessage (structured data,
Debug suffices).

Tests for the conversions + BTreeMap-key usage. Types 127 green; clippy clean;
workspace + dawai-model build.

+++ crates/midi/tutti-midi-types/src/message.rs
@@ -261,9 +261,7 @@ impl MidiEvent {
                     id: note_id(note),
                     channel,
                     note,
-                    controller: PerNoteController::Assignable {
-                        index: u8::from(m.index()),
-                    },
+                    controller: PerNoteController::Assignable { index: m.index() },
                     value: m.controller_data(),
                 }
             }
